pyproject.toml
[project]
name = "adpredictor"
version = "0.1.0"
description = "An implementation of adpredictor, a Bayesian click-through rate prediction algorithm for advertising."
readme = "README.md"
authors = [
    { name = "Ewerton de Oliveira <ewerlopes@gmail.com>" }
]
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "fonttools==4.58.4",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.3",
    "mypy>=1.16.1",
    "numpy==2.2.6",
    "packaging==25.0",
    "pillow==11.2.1",
    "pre-commit>=4.2.0",
    "protobuf==6.31.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "ruff>=0.12.1",
    "scipy==1.15.3",
    "six==1.17.0",
]
